move scalar to its own module
